+++ src/GL/Window/Window_X11.cpp
@@ -238,7 +238,7 @@ namespace GL
 		}
 	}
 
-	Context& Window::GetContext( uint color, uint depth, uint stencil, uint antialias )
+	Context& Window::GetContext( uchar color, uchar depth, uchar stencil, uchar antialias )
 	{
 		if ( context )
 			return *context;
